Rename plugins to more clear names, use the gradle plugin publishing plugin
