+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'xenit-gradle-plugins'
+rootProject.name = 'alfresco-docker-gradle-plugin'
