+++ src/integration-test/examples/alfresco-war-only/build.gradle
@@ -1,6 +1,7 @@
 plugins {
-   id  'xenit-applyamps'
+   id 'eu.xenit.docker-alfresco'
 }
+
 repositories {
    mavenCentral()
    jcenter()
