+++ src/integration-test/examples/file-dependencies/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id  'xenit-applyamps'
+    id 'eu.xenit.docker-alfresco'
 }
 
 repositories {
