+++ src/integration-test/reproductions/applyamps-subproject/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'xenit-applyamps'
+    id 'eu.xenit.docker-alfresco'
 }
 
 repositories {
