+++ src/integration-test/reproductions/different-version-properties-between-war-and-image-should-fail/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'xenit-applyamps'
+    id 'eu.xenit.docker-alfresco'
 }
 
 repositories {
