+++ src/integration-test/reproductions/docker-plugin-without-config/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'xenit-dockerbuild'
+    id 'eu.xenit.docker'
 }
 repositories {
     mavenCentral()
