+++ src/integration-test/reproductions/same-version-properties-between-war-and-image-should-succeed/build.gradle
@@ -4,7 +4,7 @@ buildscript {
     }
 }
 plugins {
-    id 'xenit-applyamps'
+    id 'eu.xenit.docker-alfresco'
 }
 
 repositories {
